+++ stroom-core-client/src/main/java/stroom/pipeline/stepping/client/presenter/SteppingPresenter.java
@@ -275,8 +275,8 @@ public class SteppingPresenter extends MyPresenterWidget<SteppingPresenter.Stepp
         // the right data.
         final SourceLocation sourceLocation = SourceLocation.builder(meta.getId())
                 .withChildStreamType(childStreamType)
-                .withPartNo(0L)
-                .withSegmentNumber(-1L)
+                .withPartIndex(stepLocation.getPartIndex())
+                .withRecordIndex(stepLocation.getRecordIndex())
                 .build();
         sourcePresenter.setSourceLocation(sourceLocation);
 
@@ -314,11 +314,11 @@ public class SteppingPresenter extends MyPresenterWidget<SteppingPresenter.Stepp
                         AlertEvent.fireError(SteppingPresenter.this, e.getMessage(), null);
                     }
 
-                    if (stepLocation != null && stepLocation.getRecordNo() > 0) {
+                    if (stepLocation.getRecordIndex() > -1) {
                         step(StepType.REFRESH, new StepLocation(
                                 meta.getId(),
-                                stepLocation.getPartNo(),
-                                stepLocation.getRecordNo()));
+                                stepLocation.getPartIndex(),
+                                stepLocation.getRecordIndex()));
                     }
                 })
                 .call(PIPELINE_RESOURCE)