+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/DataVolumeDao.java
@@ -17,7 +17,7 @@ public interface DataVolumeDao {
 
     interface DataVolume {
 
-        long getStreamId();
+        long getMetaId();
 
         String getVolumePath();
     }