+++ stroom-pipeline/src/test/java/stroom/pipeline/xml/converter/json/TestJSONParser.java
@@ -157,10 +157,10 @@ class TestJSONParser extends StroomUnitTest {
                 final ZipInputStream zipInputStream = new ZipInputStream(Files.newInputStream(input));
                 try {
                     ZipEntry entry = zipInputStream.getNextEntry();
-                    long streamNo = 0;
+                    long partIndex = -1;
                     while (entry != null) {
-                        streamNo++;
-                        locationFactory.setStreamNo(streamNo);
+                        partIndex++;
+                        locationFactory.setPartIndex(partIndex);
 
                         try {
                             if (entry.getName().endsWith(".dat")) {