+++ stroom-ui/src/api/stroom.ts
@@ -974,7 +974,7 @@ export interface FetchDataRequest {
   pipeline?: DocRef;
 
   /** @format int64 */
-  segmentCount?: number;
+  recordCount?: number;
   showAsHtml?: boolean;
   sourceLocation?: SourceLocation;
 }
@@ -2152,10 +2152,10 @@ export interface RangeLong {
 
 export interface Rec {
   /** @format int64 */
-  recordNo?: number;
+  metaId?: number;
 
   /** @format int64 */
-  streamId?: number;
+  recordIndex?: number;
 }
 
 export interface ReceiveDataRule {
@@ -2225,15 +2225,15 @@ export interface RefStoreEntry {
 }
 
 export interface RefStreamDefinition {
+  /** @format int64 */
+  partIndex?: number;
+
   /** A class for describing a unique reference to a 'document' in stroom.  A 'document' is an entity in stroom such as a data source dictionary or pipeline. */
   pipelineDocRef?: DocRef;
   pipelineVersion?: string;
 
   /** @format int64 */
   streamId?: number;
-
-  /** @format int64 */
-  streamNo?: number;
 }
 
 export interface ReferenceLoader {
@@ -2811,13 +2811,13 @@ export interface SourceLocation {
   highlight?: TextRange;
 
   /** @format int64 */
-  id?: number;
+  metaId?: number;
 
   /** @format int64 */
-  partNo?: number;
+  partIndex?: number;
 
   /** @format int64 */
-  segmentNo?: number;
+  recordIndex?: number;
   truncateToWholeLines?: boolean;
 }
 
@@ -2875,13 +2875,13 @@ export interface StatisticsDataSourceFieldChangeRequest {
 
 export interface StepLocation {
   /** @format int64 */
-  id?: number;
+  metaId?: number;
 
   /** @format int64 */
-  partNo?: number;
+  partIndex?: number;
 
   /** @format int64 */
-  recordNo?: number;
+  recordIndex?: number;
 }
 
 export interface SteppingFilterSettings {
@@ -2933,7 +2933,7 @@ export interface StoredQuery {
   version?: number;
 }
 
-export type StreamLocation = Location & { streamNo?: number };
+export type StreamLocation = Location & { partIndex?: number };
 
 export interface StringCriteria {
   caseInsensitive?: boolean;