+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2275** : Stepping from data popup now takes you to the correct record.
+
 * Issue **#2263** : Removed unnecessary JWS algorithm constraints.
 
 * Issue **#2240** : Indexes now show empty selection for volume group until one is selected.